# opt: reduce allocations in OrderingChoice simplification