--- a/pkg/sql/opt/props/ordering_choice.go
+++ b/pkg/sql/opt/props/ordering_choice.go
@@ -657,12 +657,15 @@ func (oc *OrderingChoice) AppendCol(id opt.ColumnID, descending bool) {
 }
 
 // Copy returns a complete copy of this instance, with a private version of the
-// ordering column array.
+// ordering column slice.
 func (oc *OrderingChoice) Copy() OrderingChoice {
 	var other OrderingChoice
 	other.Optional = oc.Optional.Copy()
 	other.Columns = make([]OrderingColumnChoice, len(oc.Columns))
 	copy(other.Columns, oc.Columns)
+	for i := range other.Columns {
+		other.Columns[i].Group = other.Columns[i].Group.Copy()
+	}
 	return other
 }