#### opt: move ComputeEquivGroup method

This commit moves the `FuncDep.ComputeEquivGroup` method next to other
methods that perform similar computations, making it easier to maintain
them all.

Release note: None

#### opt: reduce intset.Fast allocations in OrderingChoice.Simplify

This commit introduces the `FuncDep.ComputeClosureNoCopy` method which
is similar to `FuncDep.ComputeClosure`, but mutates the given column
set in-place. This helps avoid unnecessary allocations in
`OrderingChoice.Simplify`.

This commit also reduces copying by creating the temporary `closure`
column set with a single copy, and using `UnionWith` (which does not
copy) instead of `Union` (which copies) to add new columns to `closure`.

Release note: None

#### opt: reduce intset.Fast allocations in OrderingChoice.CanSimplify

This commit reduces allocations in `OrderingChoice.CanSimplify` by
introducing two new methods for `FuncDep`: `IsClosureEqual` and
`IsEquivClosureEqual`. These methods return whether or not the closures
will be equal to the given column set, without allocating new column
sets.

This commit also reduces allocations in `OrderingChoice.CanSimplify` by
using `ComputeClosureNoCopy` instead of `ComputeClosure`.

Release note: None

#### intsets: add KeepOne method to Fast and Sparse

This commit adds the `KeepOne` method to `Fast` and `Sparse` which
removes all but one element from the set. The method is a no-op if the
set is empty.

Release note: None

#### opt: copy ordering choice columns

This commit fixes a long-standing bug in `OrderingChoice.Copy` in which
the `Group` columns were not being copied. As far as we now, this has
caused no user-facing bugs, but it does not follow `Copy`'s description
that the new `OrderingChoice` has a "private" version of the ordering
column slice, and column sets shared between `OrderingChoice`s could
lead to future bugs.

Release note: None

#### opt: reduce allocations in OrderingChoice.Simplify with Fast.KeepOne

This commit reduces allocations in `OrderingChoice.Simplify` by reducing
a group's set with `Fast.KeepOne` and passing the set to
`FuncDep.ComputeEquivClosureNoCopy`. This avoids allocations in
`FuncDep.ComputeEquivGroup` when creating singleton sets with
`MakeColSet` when the column ID is outside the range [1, 128].

Release note: None

#### opt: reduce allocations in OrderingChoice.CanSimplify further

This commit further reduces allocations in `OrderingChoice.CanSimplify`
by using a new method `FuncDep` method, `ComputeEquivGroupAndUnionWith`.
This method avoid allocations in `ComputeEquivGroup` when building up a
temporary column set.

Epic: None

Release note (performance improvement): The optimizer now optimizes
query plans with many joins more quickly and uses significantly less
memory while doing so.