@mgartner 2023-04-11
### Optimizer benchmark results

Slow query 4 was the focus of this effort, and it is significantly faster and uses less memory:

```
name                         old time/op    new time/op    delta
SlowQueries/slow-query-1-10    15.7ms ± 1%    15.4ms ± 1%   -1.52%  (p=0.032 n=5+5)
SlowQueries/slow-query-2-10     219ms ± 0%     222ms ± 0%   +1.41%  (p=0.008 n=5+5)
SlowQueries/slow-query-3-10    62.7ms ± 0%    62.5ms ± 1%     ~     (p=0.310 n=5+5)
SlowQueries/slow-query-4-10     1.37s ± 0%     1.05s ± 0%  -23.61%  (p=0.008 n=5+5)

name                         old alloc/op   new alloc/op   delta
SlowQueries/slow-query-1-10    6.99MB ± 0%    6.58MB ± 0%   -5.74%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-10    48.7MB ± 0%    48.7MB ± 0%     ~     (p=0.421 n=5+5)
SlowQueries/slow-query-3-10    44.9MB ± 0%    44.8MB ± 0%   -0.22%  (p=0.008 n=5+5)
SlowQueries/slow-query-4-10     737MB ± 0%     373MB ± 0%  -49.37%  (p=0.016 n=5+4)

name                         old allocs/op  new allocs/op  delta
SlowQueries/slow-query-1-10     75.8k ± 0%     63.3k ± 0%  -16.52%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-10      400k ± 0%      400k ± 0%     ~     (p=0.151 n=5+5)
SlowQueries/slow-query-3-10      389k ± 0%      386k ± 0%   -0.78%  (p=0.008 n=5+5)
SlowQueries/slow-query-4-10     17.4M ± 0%      6.1M ± 0%  -65.04%  (p=0.008 n=5+5)
```

The Exploration Phase benchmarks are mostly the same. There is some improvement in `many-columns-and-indexes-c` and a slight regression in `ored-preds-100`, but I believe it is acceptable.

```
name                                                    old time/op    new time/op    delta
Phases/kv-read/Simple/Explore-10                          26.3µs ± 1%    26.1µs ± 0%     ~     (p=0.151 n=5+5)
Phases/kv-read-const/Simple/Explore-10                    27.0µs ±11%    26.4µs ± 3%     ~     (p=0.548 n=5+5)
Phases/tpcc-new-order/Simple/Explore-10                   72.5µs ± 0%    72.5µs ± 0%     ~     (p=1.000 n=5+5)
Phases/tpcc-delivery/Simple/Explore-10                    55.6µs ± 0%    55.3µs ± 0%   -0.52%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-10                  18.3µs ± 1%    18.2µs ± 2%     ~     (p=0.222 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-10                  249µs ± 0%     251µs ± 1%     ~     (p=0.095 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-10               97.8µs ± 0%    98.1µs ± 1%     ~     (p=0.548 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-10        258µs ± 0%     258µs ± 0%     ~     (p=0.690 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-10      233µs ± 1%     232µs ± 0%   -0.45%  (p=0.016 n=5+5)
Phases/many-columns-and-indexes-b/Simple/Explore-10        311µs ± 0%     310µs ± 0%     ~     (p=0.095 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-10      263µs ± 0%     264µs ± 0%     ~     (p=0.095 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-10        174ms ± 1%     171ms ± 1%   -1.69%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-10      170ms ± 2%     168ms ± 2%     ~     (p=0.548 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-10       84.8µs ± 0%    85.4µs ± 0%   +0.68%  (p=0.032 n=4+5)
Phases/single-col-histogram-range/Prepared/Explore-10     38.0µs ± 5%    37.3µs ± 1%     ~     (p=0.841 n=5+5)
Phases/batch-insert-one/Simple/Explore-10                 60.4µs ± 6%    58.7µs ± 0%   -2.78%  (p=0.008 n=5+5)
Phases/batch-insert-many/Simple/Explore-10                4.37ms ± 0%    4.47ms ± 4%     ~     (p=1.000 n=5+5)
Phases/ored-preds-100/Simple/Explore-10                   2.51ms ± 0%    2.77ms ±10%  +10.37%  (p=0.008 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-10      2.39ms ± 0%    2.40ms ± 2%     ~     (p=0.421 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-10    1.04ms ± 0%    1.04ms ± 0%   +0.39%  (p=0.008 n=5+5)

name                                                    old alloc/op   new alloc/op   delta
Phases/kv-read/Simple/Explore-10                          15.0kB ± 0%    15.0kB ± 0%     ~     (all equal)
Phases/kv-read-const/Simple/Explore-10                    15.1kB ± 0%    15.1kB ± 0%     ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-10                   32.2kB ± 0%    32.2kB ± 0%     ~     (p=0.579 n=5+5)
Phases/tpcc-delivery/Simple/Explore-10                    26.1kB ± 0%    26.1kB ± 0%     ~     (p=0.246 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-10                  12.4kB ± 0%    12.4kB ± 0%     ~     (all equal)
Phases/tpcc-stock-level/Simple/Explore-10                  128kB ± 0%     128kB ± 0%     ~     (p=0.444 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-10               56.9kB ± 0%    56.9kB ± 0%     ~     (p=0.151 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-10       39.4kB ± 0%    39.4kB ± 0%     ~     (p=0.159 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-10     5.36kB ± 0%    5.36kB ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-10       47.7kB ± 0%    47.7kB ± 0%     ~     (p=0.460 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-10     7.88kB ± 0%    7.88kB ± 0%     ~     (p=0.238 n=5+4)
Phases/many-columns-and-indexes-c/Simple/Explore-10        107MB ± 0%     105MB ± 0%   -2.54%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-10      106MB ± 0%     103MB ± 0%   -2.57%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-10       29.4kB ± 0%    29.4kB ± 0%     ~     (p=0.056 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-10     12.8kB ± 0%    12.8kB ± 0%     ~     (all equal)
Phases/batch-insert-one/Simple/Explore-10                 64.9kB ± 0%    64.9kB ± 0%     ~     (p=0.111 n=5+5)
Phases/batch-insert-many/Simple/Explore-10                3.94MB ± 0%    3.94MB ± 0%     ~     (p=0.690 n=5+5)
Phases/ored-preds-100/Simple/Explore-10                    947kB ± 0%     947kB ± 0%     ~     (p=0.690 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-10       928kB ± 0%     928kB ± 0%     ~     (p=0.206 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-10     738kB ± 0%     738kB ± 0%     ~     (p=0.452 n=5+5)

name                                                    old allocs/op  new allocs/op  delta
Phases/kv-read/Simple/Explore-10                            90.0 ± 0%      90.0 ± 0%     ~     (all equal)
Phases/kv-read-const/Simple/Explore-10                      90.0 ± 0%      90.0 ± 0%     ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-10                      169 ± 0%       169 ± 0%     ~     (all equal)
Phases/tpcc-delivery/Simple/Explore-10                       159 ± 0%       159 ± 0%     ~     (all equal)
Phases/tpcc-delivery/Prepared/Explore-10                    71.0 ± 0%      71.0 ± 0%     ~     (all equal)
Phases/tpcc-stock-level/Simple/Explore-10                    695 ± 0%       695 ± 0%     ~     (all equal)
Phases/tpcc-stock-level/Prepared/Explore-10                  331 ± 0%       331 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-a/Simple/Explore-10         76.0 ± 0%      76.0 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-a/Prepared/Explore-10       28.0 ± 0%      28.0 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-10          131 ± 0%       131 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-b/Prepared/Explore-10       46.0 ± 0%      46.0 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-c/Simple/Explore-10         734k ± 0%      649k ± 0%  -11.61%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-10       729k ± 0%      643k ± 0%  -11.71%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-10          517 ± 0%       517 ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Prepared/Explore-10        248 ± 0%       248 ± 0%     ~     (all equal)
Phases/batch-insert-one/Simple/Explore-10                    326 ± 0%       326 ± 0%     ~     (all equal)
Phases/batch-insert-many/Simple/Explore-10                 28.5k ± 0%     28.5k ± 0%     ~     (p=0.476 n=5+5)
Phases/ored-preds-100/Simple/Explore-10                    4.53k ± 0%     4.53k ± 0%     ~     (all equal)
Phases/ored-preds-using-params-100/Simple/Explore-10       4.53k ± 0%     4.53k ± 0%     ~     (all equal)
Phases/ored-preds-using-params-100/Prepared/Explore-10     2.88k ± 0%     2.88k ± 0%     ~     (all equal)
```